pyproject.toml
[project]
name = "tts-keyboard"
version = "0.1.1"
description = """
    a simple voice keyboard which uses Piper for vocal synthesis and PyAudio
    for playback.
"""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "piper-tts>=1.3.0",
    "PyAudio>=0.2.0",
    "sanitize-filename>=1.2.0",
]
